Move application version code to its own package.
This allows access to the application version info from other wallet
packages, which is of particular use for useragent strings when using
the wire protocol.
